Init
I'm picking a bit of an arbitrary point do this but things mostly work
right now so might as well.

What we have is a Cloudflare Durable object that tracks assertions of
facts, and a client that dumps those facts into replicache and renders
them.

There is a form for asserting new facts as well.