Share mutations across backend/front-end
Mutations accept an `assert` function which is implemented differently
across the front and backends. This will need to be extended to enable
querying as well in the future.

We will also need to extend it to allow for different behavior on the
client and serverside. My current assumption is this can be handled by
hooks that run before and after each mutation, keep the body of them the
same, but that might change.